+++ coaster/manage.py
@@ -24,6 +24,7 @@ def alembic_table_metadata():
 
 
 class InitedMigrations(ManageMigrations):
+    """Perform Alembic database migration operations"""
     def run(self, args):
         if len(args) and not args[0].startswith('-'):
             manager.init_for(args[0])
@@ -31,8 +32,7 @@ class InitedMigrations(ManageMigrations):
 
 
 def set_alembic_revision(path=None):
-    """Create/Update alembic table to latest revision number
-    """
+    """Create/Update alembic table to latest revision number"""
     config = Config()
     try:
         config.set_main_option("script_location", path or "alembic")
@@ -57,16 +57,16 @@ def set_alembic_revision(path=None):
 
 @database.option('-e', '--env', default='dev', help="runtime environment [default 'dev']")
 def drop(env):
-    "Drops database tables"
+    "Drop database tables"
     manager.init_for(env)
     manager.db.engine.echo = True
-    if prompt_bool("Are you sure you want to lose all your data?"):
+    if prompt_bool("Are you sure you want to lose all your data"):
         manager.db.drop_all()
 
 
 @database.option('-e', '--env', default='dev', help="runtime environment [default 'dev']")
 def create(env):
-    "Creates database tables from sqlalchemy models"
+    "Create database tables from sqlalchemy models"
     manager.init_for(env)
     manager.db.engine.echo = True
     manager.db.create_all()
@@ -75,7 +75,7 @@ def create(env):
 
 @manager.option('-e', '--env', default='dev', help="runtime environment [default 'dev']")
 def sync_resources(env):
-    "Syncs resources for the client on Lastuser server"
+    """Sync the client's resources with the Lastuser server"""
     manager.init_for(env)
     print "Syncing resources with Lastuser..."
     resources = manager.app.lastuser.sync_resources()['results']
@@ -92,14 +92,25 @@ def sync_resources(env):
                     print "\tAction %s/%s %s..." % (rname, aname, resource['status'])
     print "Resources synced..."
 
-@manager.shell
-def _make_context():
-    manager.init_for('prod')
-    return dict(app=manager.app, db=manager.db, init_for=manager.init_for, flask=flask)
+
+@manager.option('-e', '--env', default='dev', help="shell environment [default dev]")
+def shell(env, no_ipython=False, no_bpython=False):
+    """Initiate a Python shell"""
+    def _make_context():
+        manager.init_for(env)
+        return dict(app=manager.app, db=manager.db, init_for=manager.init_for, flask=flask)
+    Shell(make_context=_make_context).run(no_ipython=no_ipython, no_bpython=no_bpython)
+
+
+@manager.option('-e', '--env', default='dev', help="shell environment [default dev]")
+def plainshell(env):
+    """Initiate a plain Python shell"""
+    return shell(env, no_ipython=True, no_bpython=True)
 
 
 def init_manager(app, db, init_for, **kwargs):
-    """Initialise Manager
+    """
+    Initialise Manager
 
     :param app: Flask app object
     :parm db: db instance
@@ -110,5 +121,4 @@ def init_manager(app, db, init_for, **kwargs):
     manager.add_command("clean", Clean())
     manager.add_command("showurls", ShowUrls())
     manager.add_command("migrate", InitedMigrations())
-    manager.add_command("shell", Shell(make_context=_make_context))
     return manager